persist cosmetics data
