+++ sql/create_tables.cql
@@ -23,7 +23,8 @@ CREATE TABLE matches (
   radiant_gold_adv text,
   radiant_xp_adv text,
   teamfights text,
-  version int
+  version int,
+  cosmetics text,
 );
 
 CREATE TABLE player_matches (
