+++ sql/migrations.cql
@@ -1,0 +1,1 @@
+ALTER TABLE matches ADD cosmetics text;
